third_party/WebKit/LayoutTests/media/remoteplayback/watch-availability-initial-callback.html

<!DOCTYPE html>
<html>
    <head>
        <title>Test that the callback is called once watchAvailability() resolves.</title>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
        <script src="../media-file.js"></script>
    </head>
    <body>
        <script>
            async_test(function(t)
            {
                var v = document.createElement('video');
                v.src = findMediaFile('video', 'content/test');
                document.body.appendChild(v);

                var promiseResolved = false;

                function callback(available) {
                  assert_false(available);
                  assert_true(promiseResolved);
                }

                v.remote.watchAvailability(t.step_func_done(callback)).then(
                        function() { promiseResolved = true; },
                        t.unreached_func());
            }, 'Test that the callback is called once watchAvailability() resolves.');
        </script>
    </body>
</html>
